Changed the ServerBaseTest to also create a socket, with the appropriate streams, as all uses of the file require a socket to check communication with the server
