+++ java/tests/server/connection/SocketHandlerTest.java
@@ -2,18 +2,14 @@ package server.connection;
 
 import static junit.framework.TestCase.assertTrue;
 
-import java.io.IOException;
-import java.net.InetSocketAddress;
-import java.net.Socket;
-import org.junit.Test;
 import base.ServerBaseTest;
+import java.io.IOException;
+import org.junit.Test;
 
 public class SocketHandlerTest extends ServerBaseTest {
 
   @Test
   public void testConnection() throws IOException {
-    Socket socket = new Socket();
-    socket.connect(new InetSocketAddress("localhost", SocketHandler.PORT));
     assertTrue(socket.isConnected());
   }
